refactor: 분석 채팅 컨테이너 훅 분리
